backend/requirements.txt
asgi-lifespan==2.0.0
asgiref==3.5.2
channels-redis==3.4.1
channels==3.0.5
django-allauth==0.52.0
django-asgi-lifespan==0.1.0
django-cors-headers==3.13.0
django-feather==0.3.0
django-rest-knox==4.2.0
django-robots==5.0
django-storages==1.13.2
Django==4.1.3
djangorestframework==3.13.1
gunicorn==20.1.0
httpx==0.23.3
mysqlclient==2.1.1
pillow==9.2.0
python-magic==0.4.27
requests==2.28.1
sentry-sdk==1.14.0
six==1.16.0
tzdata==2022.7
uvicorn==0.18.2
hgtk==0.1.3
